refactoring: user to member
